[FLINK-27998][table] Address feedback for row constructor
