+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/copy/HiveParserSqlFunctionConverter.java
@@ -21,7 +21,6 @@ package org.apache.flink.table.planner.delegation.hive.copy;
 import org.apache.flink.table.planner.delegation.hive.HiveParserIN;
 import org.apache.flink.table.planner.delegation.hive.SqlFunctionConverter;
 import org.apache.flink.table.planner.delegation.hive.parse.HiveASTParser;
-import org.apache.flink.table.planner.functions.sql.FlinkSqlOperatorTable;
 
 import org.apache.calcite.rel.type.RelDataType;
 import org.apache.calcite.sql.SqlAggFunction;
@@ -190,9 +189,7 @@ public class HiveParserSqlFunctionConverter {
                     HiveParserBetween.INSTANCE,
                     hToken(HiveASTParser.Identifier, "between"));
             registerFunction(
-                    "struct",
-                    FlinkSqlOperatorTable.ROW,
-                    hToken(HiveASTParser.Identifier, "struct"));
+                    "struct", SqlStdOperatorTable.ROW, hToken(HiveASTParser.Identifier, "struct"));
             registerFunction(
                     "isnotnull",
                     SqlStdOperatorTable.IS_NOT_NULL,
